[reduce-deps] move main to q to test it properly
